preserve custom style
