+++ src/pages/Admin/Charity/Dashboard/Schedule/Edit.tsx
@@ -7,13 +7,20 @@ import { useEditEndowmentMutation } from "services/aws/aws";
 import type { Allocation } from "types/aws";
 import { AllocationOptions } from "./AllocationOptions";
 import { AllocationSlider } from "./AllocationSlider";
+import { allocationOptions, toKey } from "./common";
 
-export function Edit(props: Allocation & { id: number; amount: number }) {
+export function Edit({
+  id,
+  amount,
+  ...props
+}: Allocation & { id: number; amount: number }) {
   const { closeModal } = useModalContext();
   const [editEndow, { isLoading }] = useEditEndowmentMutation();
   const { handleError } = useErrorContext();
   const [alloc, setAlloc] = useState<Allocation>(props);
-  const [isCustom, setIsCustom] = useState(false);
+  const [isCustom, setIsCustom] = useState(
+    allocationOptions.every((opt) => opt.value !== toKey(props))
+  );
 
   return (
     <Modal className="fixed-center z-10 grid gap-y-4 text-navy-d4 dark:text-white bg-white dark:bg-blue-d4 sm:w-full w-[90vw] sm:max-w-lg rounded-lg p-6">
@@ -54,7 +61,7 @@ export function Edit(props: Allocation & { id: number; amount: number }) {
         className="btn btn-blue px-4 py-2 text-sm uppercase mt-4 rounded-full"
         onClick={async () => {
           try {
-            await editEndow({ id: props.id, allocation: alloc }).unwrap();
+            await editEndow({ id, allocation: alloc }).unwrap();
             closeModal();
           } catch (err) {
             handleError(err);
